connection can not be on init because cause a command line panic
